Upload (partial) of latest work on Component infrastructure.
Minor changes in some SFINAE to more-closely resemble certain std:: constructs.

Sfome SFINAE sfimplifications.

Removed unnecessary "inline"s. The functions in question were in fact function *templates*, so that "inline" was unnecessary in order to still achieve header-only-ness.

Used "struct" instead of "class" in a few places.

Used "size_t" instead of "int" in a few places. The difference wasn't really important in these places, but I decided that size_t was correct in principle.

C Interface: fixed some previously unnoticed discrepancies that arose after an earlier change. Some function names were changed a while back: color to colors, note to notes, and warning to warnings. However, the *declarations* (for C) hadn't been updated accordingly.

Function name change: detail::getter() to detail::compGetter(). (comp, for class Component.) Too many things were called "getter." This change should make code auditing/analysis a bit easier, going forward.

Miscellaneous other changes. There are still some relevant files to be uploaded, but I need to review them more carefully first.

+++ src/GNDStk/HDF5.hpp
@@ -31,8 +31,8 @@ public:
    // temporaryName()
    static std::string temporaryName(int &desc)
    {
-      static const int maxTry = 10;
-      int count = 0;
+      static const size_t maxTry = 10;
+      size_t count = 0;
 
       while (true) {
          std::string name = "GNDStk-tmp-XXXXXX.h5";
